[doc] Begin to improve the Upgrade Guide based on #1140.
