+++ examples/format-dynamic-content.js
