+++ examples/format-logger-and-transport.js
