+++ examples/metadata.js
@@ -5,6 +5,10 @@ const winston = require('../');
 const logger = winston.createLogger({
   level: 'info',
   format: winston.format.combine(
+    //
+    // Notice that both arguments have been combined into a single
+    // "info" object.
+    //
     winston.format(function (info, opts) {
       console.log(`{ reason: ${info.reason}, promise: ${info.promise} }`);
       return info;
